Fix staging timer

+++ Core/Models/DevelopmentParameters.swift
@@ -12,10 +12,10 @@ struct DevelopmentParameters {
     let film: SwiftDataFilm
     let developer: SwiftDataDeveloper
     let dilution: String
-    let temperature: Double
+    let temperature: Int
     let iso: Int
     
-    init(film: SwiftDataFilm, developer: SwiftDataDeveloper, dilution: String, temperature: Double, iso: Int) {
+    init(film: SwiftDataFilm, developer: SwiftDataDeveloper, dilution: String, temperature: Int, iso: Int) {
         self.film = film
         self.developer = developer
         self.dilution = dilution
