+++ Core/Models/SwiftDataDevelopmentParameters.swift
@@ -11,10 +11,10 @@ struct SwiftDataDevelopmentParameters {
     let film: SwiftDataFilm
     let developer: SwiftDataDeveloper
     let dilution: String
-    let temperature: Double
+    let temperature: Int
     let iso: Int
     
-    init(film: SwiftDataFilm, developer: SwiftDataDeveloper, dilution: String, temperature: Double, iso: Int) {
+    init(film: SwiftDataFilm, developer: SwiftDataDeveloper, dilution: String, temperature: Int, iso: Int) {
         self.film = film
         self.developer = developer
         self.dilution = dilution
